moved mapview out of activity
